tests/requirements-test.txt
pytest==7.4.3
pytest-cov==4.1.0
pytest-mock==3.12.0
pytest-xdist==3.3.1
coverage==7.3.2
pytest-flask==1.3.0 